Collected SM metrics from (DQN, atari). Experimenting with running inference in TensorRT.

+++ stable_baselines/common/input.py
@@ -22,8 +22,10 @@ def observation_input(ob_space, batch_size=None, name='Ob', scale=False):
         return observation_ph, processed_observations
 
     elif isinstance(ob_space, Box):
-        observation_ph = tf.compat.v1.placeholder(shape=(batch_size,) + ob_space.shape, dtype=ob_space.dtype, name=name)
-        processed_observations = tf.cast(observation_ph, tf.float32)
+        # observation_ph = tf.compat.v1.placeholder(shape=(batch_size,) + ob_space.shape, dtype=ob_space.dtype, name=name)
+        observation_ph = tf.compat.v1.placeholder(shape=(batch_size,) + ob_space.shape, dtype=tf.float32, name=name)
+        # processed_observations = tf.cast(observation_ph, tf.float32) # TRT CALLED (twice)
+        processed_observations = observation_ph
         # rescale to [1, 0] if the bounds are defined
         if (scale and
            not np.any(np.isinf(ob_space.low)) and not np.any(np.isinf(ob_space.high)) and
